requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
anthropic==0.43.0
pydantic==2.10.4
apscheduler==3.10.4
python-dotenv==1.0.1
jsonlines==4.0.0
jinja2==3.1.5
